fix: lib64 builds fail once on clean install
Due to `find_node` creating a task, its not guaranteed to run after the
completion of CMakeBuildTask. Since we depended on the output of
CMakeBuildTask to switch to lib64, the first build would tend to return
false and incorrectly give `lib` on `lib64` based systems.

The fix is to read both paths in read_stlib, instead of using an if
statement to choose the correct lib folder.

Fixes: steinwurf/srt#3